construindo api com laravel

+++ app/Http/Controllers/ImageController.php
@@ -4,7 +4,6 @@ namespace App\Http\Controllers;
 
 use App\User;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Storage;
 use Image;
 
@@ -77,11 +76,17 @@ class ImageController extends Controller
     }
 
     
-
     public function excluir(Request $request){
         $foto = $request->get('image');
-        Storage::delete('public/img/normal/'.$foto);
-        Storage::delete('public/img/thumbnail/'.'_small_'.$foto);
+        if ($foto != 'boy.png'){
+            Storage::delete('public/img/normal/'.$foto);
+            Storage::delete('public/img/thumbnail/'.'_small_'.$foto);
+            if ($request->id){
+                $usuario = User::find($request->id);
+                $usuario->profile_pic = 'boy.png';
+                $usuario->save();
+            }
+        }
         return response()->json(array('nomeArquivo'=>'boy.png'));  
     }
